requirements.txt
altgraph==0.17.5
certifi==2025.11.12
charset-normalizer==3.4.4
customtkinter==5.2.2
darkdetect==0.8.0
idna==3.11
packaging==25.0
pefile==2024.8.26
pyinstaller==6.17.0
pyinstaller-hooks-contrib==2025.11
python-dotenv==1.2.1
pywin32-ctypes==0.2.3
requests==2.31.0
setuptools==80.9.0
urllib3==2.6.2
